--- src/main/java/com/travelland/repository/TripRepository.java
@@ -1,7 +1,0 @@
-package com.travelland.repository;
-
-import com.travelland.domain.Trip;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface TripRepository extends JpaRepository<Trip, Long> {
-}
