+++ src/main/java/com/travelland/repository/trip/TripRepository.java
@@ -1,0 +1,8 @@
+package com.travelland.repository.trip;
+
+import com.travelland.domain.Trip;
+import com.travelland.repository.trip.querydsl.CustomTripRepositoryV2;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface TripRepository extends JpaRepository<Trip, Long>, CustomTripRepositoryV2 {
+}
