+++ src/main/java/com/travelland/service/trip/TripScrapService.java
@@ -1,21 +1,19 @@
-package com.travelland.service;
+package com.travelland.service.trip;
 
-import com.querydsl.jpa.impl.JPAQueryFactory;
-import com.travelland.domain.*;
+import com.travelland.domain.Member;
+import com.travelland.domain.Trip;
+import com.travelland.domain.TripScrap;
 import com.travelland.dto.TripDto;
 import com.travelland.global.exception.CustomException;
 import com.travelland.global.exception.ErrorCode;
 import com.travelland.repository.MemberRepository;
-import com.travelland.repository.TripRepository;
-import com.travelland.repository.TripScrapRepository;
+import com.travelland.repository.trip.TripRepository;
+import com.travelland.repository.trip.TripScrapRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
-import java.util.stream.Collectors;
-
-import static com.travelland.domain.QTripScrap.*;
 
 @Service
 @RequiredArgsConstructor
@@ -24,7 +22,7 @@ public class TripScrapService {
     private final TripScrapRepository tripScrapRepository;
     private final MemberRepository memberRepository;
     private final TripRepository tripRepository;
-    private final JPAQueryFactory jpaQueryFactory;
+
 
     //여행정보 스크랩 등록
     @Transactional
@@ -52,18 +50,9 @@ public class TripScrapService {
     
     //스크랩한 여행정보 목록 조회
     @Transactional(readOnly = true)
-    public List<TripDto.GetTripScrapListResponse> getTripScrapList(int page, int size, String email) {
-        Member member = getMember(email);
-
-        List<TripScrap> tripScrapList = jpaQueryFactory.selectFrom(tripScrap)
-                .where(tripScrap.member.eq(member), tripScrap.isDeleted.eq(false))
-                .orderBy(tripScrap.trip.createdAt.desc())
-                .limit(size)
-                .offset((long) (page - 1) * size)
-                .fetch();
-
-        return tripScrapList.stream().map(tripScrap -> new TripDto.GetTripScrapListResponse(tripScrap.getTrip(), tripScrap.getMember()))
-                .collect(Collectors.toList());
+    public List<TripDto.Scraps> getTripScrapList(int page, int size, String email) {
+        return tripScrapRepository.getScrapListByMember(getMember(email),size, page)
+                .stream().map(TripDto.Scraps::new).toList();
     }
 
     //스크랩 데이터 삭제
